requirements.txt
dbfread
pandas
numpy
geopandas
shapely
geopy
scipy
networkx
pymetis
seaborn
tqdm
folium
matplotlib
pyyaml
scikit-learn
joblib
